job_search_platform/job_search_platform/requirements.txt
anyio==4.0.0
bcrypt==4.0.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
faunadb==4.5.1
Flask==3.0.0
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.7
Werkzeug==3.0.0
